created controllers file and properly put routes in there to shorten server file
